screenshot showing nmap scan  logged.
This screenshot was taken after i run a basic nmap scan on the target machine and had it appear on the dashboards. it's a bit messed up here because I've not created an actual dashboard to visualize it better. You can see the tcp scans that got logged by suricata.